+++ launchdarkly/resource_launchdarkly_webhook_test.go
@@ -15,25 +15,47 @@ resource "launchdarkly_webhook" "test" {
 	name    = "example-webhook"
 	url     = "http://webhooks.com"
 	tags    = [ "terraform" ]
+	on      = true
+}	
+`
+
+	testAccWebhookCreateWithEnabled = `
+resource "launchdarkly_webhook" "test" {
+	name    = "example-webhook"
+	url     = "http://webhooks.com"
+	tags    = [ "terraform" ]
 	enabled = true
 }	
 `
 
 	testAccWebhookUpdate = `
 resource "launchdarkly_webhook" "test" {
-	name = "Example Webhook"
-	url = "http://webhooks.com/updatedUrl"
-	tags = [ "terraform", "updated" ]
-	enabled = false
+	name   = "Example Webhook"
+	url    = "http://webhooks.com/updatedUrl"
+	tags   = [ "terraform", "updated" ]
+	on     = false
 	secret = "SuperSecret"
 }
 `
 
+	testAccWebhookWithStatements = `
+resource "launchdarkly_webhook" "with_statements" {
+	name    = "Webhook with policy statements"
+	url     = "http://webhooks.com"
+	on      = true
+	statements {
+		actions   = ["*"]	
+		effect    = "allow"
+		resources = ["proj/*:env/production:flag/*"]
+	}
+}
+`
+	// policy_statements is deprecated but we will still support it in v1
 	testAccWebhookWithPolicyStatements = `
 resource "launchdarkly_webhook" "with_statements" {
 	name    = "Webhook with policy statements"
 	url     = "http://webhooks.com"
-	enabled = true
+	on      = true
 	policy_statements {
 		actions   = ["*"]	
 		effect    = "allow"
@@ -42,17 +64,17 @@ resource "launchdarkly_webhook" "with_statements" {
 }
 `
 
-	testAccWebhookWithPolicyStatementsUpdate = `
+	testAccWebhookWithPolicyUpdate = `
 resource "launchdarkly_webhook" "with_statements" {
 	name    = "Webhook with policy statements"
 	url     = "http://webhooks.com"
-	enabled = true
-	policy_statements {
+	on      = true
+	statements {
 		actions   = ["*"]
 		effect    = "allow"
 		resources = ["proj/test:env/production:flag/*"]
 	}
-	policy_statements {
+	statements {
 		actions   = ["*"]
 		effect    = "allow"
 		resources = ["proj/test:env/production:segment/*"]
@@ -64,8 +86,8 @@ resource "launchdarkly_webhook" "with_statements" {
 resource "launchdarkly_webhook" "with_statements" {
 	name    = "Webhook with policy statements"
 	url     = "http://webhooks.com"
-	enabled = true
-	policy_statements {
+	on      = true
+	statements {
 		actions   = ["*"]
 		not_actions = ["*"]
 		effect    = "allow"
@@ -89,7 +111,7 @@ func TestAccWebhook_Create(t *testing.T) {
 					testAccCheckWebhookExists(resourceName),
 					resource.TestCheckResourceAttr(resourceName, "name", "example-webhook"),
 					resource.TestCheckResourceAttr(resourceName, "url", "http://webhooks.com"),
-					resource.TestCheckResourceAttr(resourceName, "enabled", "true"),
+					resource.TestCheckResourceAttr(resourceName, "on", "true"),
 					resource.TestCheckResourceAttr(resourceName, "tags.#", "1"),
 					resource.TestCheckResourceAttr(resourceName, testAccTagKey("terraform"), "terraform"),
 					resource.TestCheckResourceAttr(resourceName, "policy_statements.#", "0"),
@@ -104,7 +126,7 @@ func TestAccWebhook_Create(t *testing.T) {
 	})
 }
 
-func TestAccWebhook_Update(t *testing.T) {
+func TestAccWebhook_CreateWithEnabled(t *testing.T) {
 	resourceName := "launchdarkly_webhook.test"
 	resource.ParallelTest(t, resource.TestCase{
 		PreCheck: func() {
@@ -113,7 +135,7 @@ func TestAccWebhook_Update(t *testing.T) {
 		Providers: testAccProviders,
 		Steps: []resource.TestStep{
 			{
-				Config: testAccWebhookCreate,
+				Config: testAccWebhookCreateWithEnabled,
 				Check: resource.ComposeTestCheckFunc(
 					testAccCheckWebhookExists(resourceName),
 					resource.TestCheckResourceAttr(resourceName, "name", "example-webhook"),
@@ -125,18 +147,62 @@ func TestAccWebhook_Update(t *testing.T) {
 				),
 			},
 			{
+				ResourceName: resourceName,
+				ImportState:  true,
+			},
+		},
+	})
+}
+
+func TestAccWebhook_Update(t *testing.T) {
+	resourceName := "launchdarkly_webhook.test"
+	resource.ParallelTest(t, resource.TestCase{
+		PreCheck: func() {
+			testAccPreCheck(t)
+		},
+		Providers: testAccProviders,
+		Steps: []resource.TestStep{
+			{
+				Config: testAccWebhookCreateWithEnabled,
+				Check: resource.ComposeTestCheckFunc(
+					testAccCheckWebhookExists(resourceName),
+					resource.TestCheckResourceAttr(resourceName, "name", "example-webhook"),
+					resource.TestCheckResourceAttr(resourceName, "url", "http://webhooks.com"),
+					resource.TestCheckResourceAttr(resourceName, "enabled", "true"),
+					resource.TestCheckResourceAttr(resourceName, "tags.#", "1"),
+					resource.TestCheckResourceAttr(resourceName, testAccTagKey("terraform"), "terraform"),
+					resource.TestCheckResourceAttr(resourceName, "policy_statements.#", "0"),
+				),
+			},
+			{
+				Config: testAccWebhookCreate,
+				Check: resource.ComposeTestCheckFunc(
+					testAccCheckWebhookExists(resourceName),
+					resource.TestCheckResourceAttr(resourceName, "name", "example-webhook"),
+					resource.TestCheckResourceAttr(resourceName, "url", "http://webhooks.com"),
+					resource.TestCheckResourceAttr(resourceName, "on", "true"),
+					resource.TestCheckResourceAttr(resourceName, "tags.#", "1"),
+					resource.TestCheckResourceAttr(resourceName, testAccTagKey("terraform"), "terraform"),
+					resource.TestCheckResourceAttr(resourceName, "policy_statements.#", "0"),
+				),
+			},
+			{
 				Config: testAccWebhookUpdate,
 				Check: resource.ComposeTestCheckFunc(
 					testAccCheckWebhookExists(resourceName),
 					resource.TestCheckResourceAttr(resourceName, "name", "Example Webhook"),
 					resource.TestCheckResourceAttr(resourceName, "url", "http://webhooks.com/updatedUrl"),
-					resource.TestCheckResourceAttr(resourceName, "enabled", "false"),
+					resource.TestCheckResourceAttr(resourceName, "on", "false"),
 					resource.TestCheckResourceAttr(resourceName, "tags.#", "2"),
 					resource.TestCheckResourceAttr(resourceName, testAccTagKey("terraform"), "terraform"),
 					resource.TestCheckResourceAttr(resourceName, testAccTagKey("updated"), "updated"),
 					resource.TestCheckResourceAttr(resourceName, SECRET, "SuperSecret"),
 					resource.TestCheckResourceAttr(resourceName, "policy_statements.#", "0"),
 				),
+			},
+			{
+				ResourceName: resourceName,
+				ImportState:  true,
 			},
 		},
 	})
@@ -151,12 +217,44 @@ func TestAccWebhook_CreateWithStatements(t *testing.T) {
 		Providers: testAccProviders,
 		Steps: []resource.TestStep{
 			{
+				Config: testAccWebhookWithStatements,
+				Check: resource.ComposeTestCheckFunc(
+					testAccCheckWebhookExists(resourceName),
+					resource.TestCheckResourceAttr(resourceName, "name", "Webhook with policy statements"),
+					resource.TestCheckResourceAttr(resourceName, "url", "http://webhooks.com"),
+					resource.TestCheckResourceAttr(resourceName, "on", "true"),
+					resource.TestCheckResourceAttr(resourceName, "statements.#", "1"),
+					resource.TestCheckResourceAttr(resourceName, "statements.0.effect", "allow"),
+					resource.TestCheckResourceAttr(resourceName, "statements.0.actions.#", "1"),
+					resource.TestCheckResourceAttr(resourceName, "statements.0.actions.0", "*"),
+					resource.TestCheckResourceAttr(resourceName, "statements.0.resources.#", "1"),
+					resource.TestCheckResourceAttr(resourceName, "statements.0.resources.0", "proj/*:env/production:flag/*"),
+				),
+			},
+			{
+				ResourceName:      resourceName,
+				ImportState:       true,
+				ImportStateVerify: true,
+			},
+		},
+	})
+}
+
+func TestAccWebhook_CreateWithPolicyStatements(t *testing.T) {
+	resourceName := "launchdarkly_webhook.with_statements"
+	resource.ParallelTest(t, resource.TestCase{
+		PreCheck: func() {
+			testAccPreCheck(t)
+		},
+		Providers: testAccProviders,
+		Steps: []resource.TestStep{
+			{
 				Config: testAccWebhookWithPolicyStatements,
 				Check: resource.ComposeTestCheckFunc(
 					testAccCheckWebhookExists(resourceName),
 					resource.TestCheckResourceAttr(resourceName, "name", "Webhook with policy statements"),
 					resource.TestCheckResourceAttr(resourceName, "url", "http://webhooks.com"),
-					resource.TestCheckResourceAttr(resourceName, "enabled", "true"),
+					resource.TestCheckResourceAttr(resourceName, "on", "true"),
 					resource.TestCheckResourceAttr(resourceName, "policy_statements.#", "1"),
 					resource.TestCheckResourceAttr(resourceName, "policy_statements.0.effect", "allow"),
 					resource.TestCheckResourceAttr(resourceName, "policy_statements.0.actions.#", "1"),
@@ -178,39 +276,44 @@ func TestAccWebhook_UpdateWithStatements(t *testing.T) {
 		Providers: testAccProviders,
 		Steps: []resource.TestStep{
 			{
-				Config: testAccWebhookWithPolicyStatements,
+				Config: testAccWebhookWithStatements,
 				Check: resource.ComposeTestCheckFunc(
 					testAccCheckWebhookExists(resourceName),
 					resource.TestCheckResourceAttr(resourceName, "name", "Webhook with policy statements"),
 					resource.TestCheckResourceAttr(resourceName, "url", "http://webhooks.com"),
-					resource.TestCheckResourceAttr(resourceName, "enabled", "true"),
-					resource.TestCheckResourceAttr(resourceName, "policy_statements.#", "1"),
-					resource.TestCheckResourceAttr(resourceName, "policy_statements.0.effect", "allow"),
-					resource.TestCheckResourceAttr(resourceName, "policy_statements.0.actions.#", "1"),
-					resource.TestCheckResourceAttr(resourceName, "policy_statements.0.actions.0", "*"),
-					resource.TestCheckResourceAttr(resourceName, "policy_statements.0.resources.#", "1"),
-					resource.TestCheckResourceAttr(resourceName, "policy_statements.0.resources.0", "proj/*:env/production:flag/*"),
+					resource.TestCheckResourceAttr(resourceName, "on", "true"),
+					resource.TestCheckResourceAttr(resourceName, "statements.#", "1"),
+					resource.TestCheckResourceAttr(resourceName, "statements.0.effect", "allow"),
+					resource.TestCheckResourceAttr(resourceName, "statements.0.actions.#", "1"),
+					resource.TestCheckResourceAttr(resourceName, "statements.0.actions.0", "*"),
+					resource.TestCheckResourceAttr(resourceName, "statements.0.resources.#", "1"),
+					resource.TestCheckResourceAttr(resourceName, "statements.0.resources.0", "proj/*:env/production:flag/*"),
 				),
 			},
 			{
-				Config: testAccWebhookWithPolicyStatementsUpdate,
+				Config: testAccWebhookWithPolicyUpdate,
 				Check: resource.ComposeTestCheckFunc(
 					testAccCheckWebhookExists(resourceName),
 					resource.TestCheckResourceAttr(resourceName, "name", "Webhook with policy statements"),
 					resource.TestCheckResourceAttr(resourceName, "url", "http://webhooks.com"),
-					resource.TestCheckResourceAttr(resourceName, "enabled", "true"),
-					resource.TestCheckResourceAttr(resourceName, "policy_statements.#", "2"),
-					resource.TestCheckResourceAttr(resourceName, "policy_statements.0.effect", "allow"),
-					resource.TestCheckResourceAttr(resourceName, "policy_statements.0.actions.#", "1"),
-					resource.TestCheckResourceAttr(resourceName, "policy_statements.0.actions.0", "*"),
-					resource.TestCheckResourceAttr(resourceName, "policy_statements.0.resources.#", "1"),
-					resource.TestCheckResourceAttr(resourceName, "policy_statements.0.resources.0", "proj/test:env/production:flag/*"),
-					resource.TestCheckResourceAttr(resourceName, "policy_statements.1.effect", "allow"),
-					resource.TestCheckResourceAttr(resourceName, "policy_statements.1.actions.#", "1"),
-					resource.TestCheckResourceAttr(resourceName, "policy_statements.1.actions.0", "*"),
-					resource.TestCheckResourceAttr(resourceName, "policy_statements.1.resources.#", "1"),
-					resource.TestCheckResourceAttr(resourceName, "policy_statements.1.resources.0", "proj/test:env/production:segment/*"),
+					resource.TestCheckResourceAttr(resourceName, "on", "true"),
+					resource.TestCheckResourceAttr(resourceName, "statements.#", "2"),
+					resource.TestCheckResourceAttr(resourceName, "statements.0.effect", "allow"),
+					resource.TestCheckResourceAttr(resourceName, "statements.0.actions.#", "1"),
+					resource.TestCheckResourceAttr(resourceName, "statements.0.actions.0", "*"),
+					resource.TestCheckResourceAttr(resourceName, "statements.0.resources.#", "1"),
+					resource.TestCheckResourceAttr(resourceName, "statements.0.resources.0", "proj/test:env/production:flag/*"),
+					resource.TestCheckResourceAttr(resourceName, "statements.1.effect", "allow"),
+					resource.TestCheckResourceAttr(resourceName, "statements.1.actions.#", "1"),
+					resource.TestCheckResourceAttr(resourceName, "statements.1.actions.0", "*"),
+					resource.TestCheckResourceAttr(resourceName, "statements.1.resources.#", "1"),
+					resource.TestCheckResourceAttr(resourceName, "statements.1.resources.0", "proj/test:env/production:segment/*"),
 				),
+			},
+			{
+				ResourceName:      resourceName,
+				ImportState:       true,
+				ImportStateVerify: true,
 			},
 		},
 	})
@@ -225,7 +328,7 @@ func TestAccWebhook_InvalidStatements(t *testing.T) {
 		Steps: []resource.TestStep{
 			{
 				Config:      testAccWebhookInvalidStatements,
-				ExpectError: regexp.MustCompile("policy_statements cannot contain both 'actions' and 'not_actions'"),
+				ExpectError: regexp.MustCompile("policy statements cannot contain both 'actions' and 'not_actions'"),
 			},
 		},
 	})